(!) Entire engine now works with float16 values instead of float32 values

+++ shaders/texture/additional_renderers/bloom_rendering_2d-Unity.shader
@@ -26,14 +26,14 @@ Shader "Custom/bloom_rendering_2d"
 
             struct appdata
             {
-                float4 vertex : POSITION;
-                float2 uv : TEXCOORD0;
+                half4 vertex : POSITION;
+                half2 uv : TEXCOORD0;
             };
 
             struct v2f
             {
-                float2 uv : TEXCOORD0;
-                float4 pos : SV_POSITION;
+                half2 uv : TEXCOORD0;
+                half4 pos : SV_POSITION;
             };
 
             v2f vert(appdata v)
@@ -53,41 +53,41 @@ Shader "Custom/bloom_rendering_2d"
                 return index;
             }
 
-            void collect_corner_colors(float3 bloom_lattice_position, out float3 colors[4])
+            void collect_corner_colors(half3 bloom_lattice_position, out half3 colors[4])
             {
-                float3 bloom_lattice_size = float3((simulation_width + 3) / 4, (simulation_height + 3) / 4, (simulation_depth + 3) / 4);
-                colors[0] = float3(0, 0, 0);
-                colors[1] = float3(0, 0, 0);
-                colors[2] = float3(0, 0, 0);
-                colors[3] = float3(0, 0, 0);
-                float3 floor_position = floor(bloom_lattice_position);
-                floor_position = clamp(floor_position, float3(0, 0, 0), bloom_lattice_size - 1);
-                float3 ceil_position = ceil(bloom_lattice_position);
-                ceil_position = clamp(ceil_position, float3(0, 0, 0), bloom_lattice_size - 1);
+                half3 bloom_lattice_size = half3((simulation_width + 3) / 4, (simulation_height + 3) / 4, (simulation_depth + 3) / 4);
+                colors[0] = half3(0, 0, 0);
+                colors[1] = half3(0, 0, 0);
+                colors[2] = half3(0, 0, 0);
+                colors[3] = half3(0, 0, 0);
+                half3 floor_position = floor(bloom_lattice_position);
+                floor_position = clamp(floor_position, half3(0, 0, 0), bloom_lattice_size - 1);
+                half3 ceil_position = ceil(bloom_lattice_position);
+                ceil_position = clamp(ceil_position, half3(0, 0, 0), bloom_lattice_size - 1);
                 uint lattice_index = 0;
                 for (uint color_channel = 0; color_channel < 3; color_channel++)
                 {
                     lattice_index = get_bloom_buffer_index(uint3(floor_position.x, floor_position.y, floor_position.z), color_channel);
-                    colors[0][color_channel] = (float)bloom_lattice_buffer[lattice_index] / 255.0f;
+                    colors[0][color_channel] = (half)bloom_lattice_buffer[lattice_index] / 255.0f;
                     lattice_index = get_bloom_buffer_index(uint3(floor_position.x, ceil_position.y, floor_position.z), color_channel);
-                    colors[1][color_channel] = (float)bloom_lattice_buffer[lattice_index] / 255.0f;
+                    colors[1][color_channel] = (half)bloom_lattice_buffer[lattice_index] / 255.0f;
                     lattice_index = get_bloom_buffer_index(uint3(ceil_position.x, floor_position.y, floor_position.z), color_channel);
-                    colors[2][color_channel] = (float)bloom_lattice_buffer[lattice_index] / 255.0f;
+                    colors[2][color_channel] = (half)bloom_lattice_buffer[lattice_index] / 255.0f;
                     lattice_index = get_bloom_buffer_index(uint3(ceil_position.x, ceil_position.y, floor_position.z), color_channel);
-                    colors[3][color_channel] = (float)bloom_lattice_buffer[lattice_index] / 255.0f;
+                    colors[3][color_channel] = (half)bloom_lattice_buffer[lattice_index] / 255.0f;
                 }
             }
 
-            float brightness = 1.0;
-            float opacity = 1.0;
+            half brightness = 1.0;
+            half opacity = 1.0;
 
-            float4 frag(v2f i) : SV_Target
+            half4 frag(v2f i) : SV_Target
             {
                 brightness = brightness ? brightness : 1.0;
                 opacity = opacity ? opacity : 1.0;
-                float3 position = float3(i.uv.x * (float)simulation_width / 4.0f, i.uv.y * (float)simulation_height / 4.0f, 0);
-                float4 color = float4(0, 0, 0, 1);
-                float3 corner_colors[4];
+                half3 position = half3(i.uv.x * (half)simulation_width / 4.0f, i.uv.y * (half)simulation_height / 4.0f, 0);
+                half4 color = half4(0, 0, 0, 1);
+                half3 corner_colors[4];
                 collect_corner_colors(position, corner_colors);
                 CommonMath::interpolate_2d(position - floor(position), corner_colors, color.xyz);
                 color *= simulation_brightness;
