+++ src/core/ops/pauli_spinor_ops.hlsl
@@ -9,8 +9,8 @@ namespace PauliSpinorOps
     // Empty (zero) a Pauli spinor
     void empty(out PauliSpinor spinor)
     {
-        spinor[0] = float2(0, 0);
-        spinor[1] = float2(0, 0);
+        spinor[0] = half2(0, 0);
+        spinor[1] = half2(0, 0);
     }
 }
